+++ test/expression/parse.test.js
@@ -430,11 +430,11 @@ describe('parse', function() {
     it('should parse function assignments', function() {
       var scope = {};
       parseAndEval('x=100', scope); // for testing scoping of the function variables
-      assert.equal(parseAndEval('function f(x) = x^2', scope).syntax, 'f(x)');
+      assert.equal(parseAndEval('f(x) = x^2', scope).syntax, 'f(x)');
       assert.equal(parseAndEval('f(3)', scope), 9);
       assert.equal(scope.f(3), 9);
       assert.equal(scope.x, 100);
-      assert.equal(parseAndEval('function g(x, y) = x^y', scope).syntax, 'g(x, y)');
+      assert.equal(parseAndEval('g(x, y) = x^y', scope).syntax, 'g(x, y)');
       assert.equal(parseAndEval('g(4,5)', scope), 1024);
       assert.equal(scope.g(4,5), 1024);
     });
@@ -442,23 +442,30 @@ describe('parse', function() {
     it ('should correctly evaluate variables in assigned functions', function () {
       var scope = {};
       assert.equal(parseAndEval('a = 3', scope), 3);
-      assert.equal(parseAndEval('function f(x) = a * x', scope).syntax, 'f(x)');
+      assert.equal(parseAndEval('f(x) = a * x', scope).syntax, 'f(x)');
       assert.equal(parseAndEval('f(2)', scope), 6);
       assert.equal(parseAndEval('a = 5', scope), 5);
       assert.equal(parseAndEval('f(2)', scope), 10);
-      assert.equal(parseAndEval('function g(x) = x^q', scope).syntax, 'g(x)');
+      assert.equal(parseAndEval('g(x) = x^q', scope).syntax, 'g(x)');
       assert.equal(parseAndEval('q = 4/2', scope), 2);
       assert.equal(parseAndEval('g(3)', scope), 9);
     });
 
     it('should throw an error for undefined variables in an assigned function', function() {
       var scope = {};
-      assert.equal(parseAndEval('function g(x) = x^q', scope).syntax, 'g(x)');
+      assert.equal(parseAndEval('g(x) = x^q', scope).syntax, 'g(x)');
       assert.throws(function () {
         parseAndEval('g(3)', scope);
       }, function (err) {
         return (err instanceof Error) && (err.toString() == 'Error: Undefined symbol q');
       });
+    });
+
+    it('should throw an error on invalid left hand side of a function assignment', function() {
+      assert.throws(function () {
+        var scope = {};
+        parseAndEval('g(x, 2) = x^2', scope);
+      }, SyntaxError);
     });
   });
 
@@ -754,7 +761,7 @@ describe('parse', function() {
       };
       assert.deepEqual(math.parse('a*b').compile(math).eval(scope), 12);
       assert.deepEqual(math.parse('c=5').compile(math).eval(scope), 5);
-      assert.deepEqual(math.parse('function f(x) = x^a').compile(math).eval(scope).syntax, 'f(x)');
+      assert.deepEqual(math.parse('f(x) = x^a').compile(math).eval(scope).syntax, 'f(x)');
 
 
       assert.deepEqual(Object.keys(scope).length, 5);